requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
flask-session==0.8.0
openai==1.16.1
pytest==8.2.0